feat: 콘텐츠 등록 위자드 단일 스텝으로 통합
- Step1에 파일 업로드/외부 링크 입력 기능 통합
- Step2(파일 업로드), Step3(설정 및 발행) 삭제
- completionCriteria: BUTTON_CLICK 고정
- downloadable: false 고정

Closes #400
