+++ src/components/domain/tu/content/index.ts
@@ -3,6 +3,4 @@ export type { ContentCardLabels } from './ContentCard';
 export { contentTypeBadgeColor, contentTypeIcon, formatFileSize, formatDate } from './ContentCard';
 export { ContentRegistrationWizard } from './ContentRegistrationWizard';
 export { Step1ContentDefinition } from './Step1ContentDefinition';
-export { Step2ContentUpload } from './Step2ContentUpload';
-export { Step3Settings } from './Step3Settings';
 export { ContentPreviewModal } from './ContentPreviewModal';
